Change dye "solving" method
Initial method was not only apparently wrong, but also stupid

+++ Export/SupabaseExporter/SupabaseExporter/Structures/Exports/Fashion.cs
@@ -27,4 +27,4 @@ public class Fashion
     }
 }
 
-public record DyeData(long Count, float Pct);
+public record DyeData(long Count, double Pct);
